Specific field CustomMsg to distinguish config reload or core dump. (#9840)
Description of PR
We have a field CustomMsg to record the result of fixture core_dump_and_config_check in DB. But previously, we record the result of core dump check and config check combinationly, either core dump check fails or config check fails, we will record false in this filed. In this PR, we will specific this filed, and record this two type of results respectively. And the format of this filed in DB is
```
{
    'DutChekResult': {
        'core_dump_check_pass': core_dump_check_pass,
        'config_db_check_pass': config_db_check_pass
    }
}
```

What is the motivation for this PR?
We have a field CustomMsg to record the result of fixture core_dump_and_config_check in DB. But previously, we record the result of core dump check and config check combinationly, either core dump check fails or config check fails, we will record false in this filed. In this PR, we will specific this filed, and record this two type of results respectively. And the format of this filed in DB is
```
{
    'DutChekResult': {
        'core_dump_check_pass': core_dump_check_pass,
        'config_db_check_pass': config_db_check_pass
    }
}
```

How did you do it?
Add more information into field CustomMsg.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>