Skip condition to skip SNMP IPv6 test case in 202211 branch. (#9856)
What is the motivation for this PR?
#9630 was added skip IPv6 related SNMP test cases in 202211 branch.
But the condition checks for branch name instead of release.
This is causing PR checker to fail as the branch name in PR checker is the PR number and not 202211

How did you do it?
Change condition to look for 202211 in release and not branch.

Signed-off-by: yenlu <yenlu@cisco.com>