Cisco: Fixing get_multiple_flows for T0 (#9362)
Addressing the following issues:

In T0 topology, since the packet can take any port in a Vlan, the get_multiple_flows function runs in to infinite loop. To avoid this, we set a limit on the loop to 20, and if it runs out, we use the port that packet is actually taking instead of the user-provided one.
addressing the keyerror in BufferPoolWatermark testcase.
Addressed the assert error in AllPortQWatermark for T0.