[platform_tests/sfp]: Fix stop_xcvrd() fixture in test_check_sfputil_reset testcase (#9651)
What is the motivation for this PR?
The test "test_check_sfputil_reset " has a fixture stop_xcvrd(). This fixture alters the file pmon_daemon_control.json to add services which should be excluded for restart. This file is not mandatory for all platforms and may not be present in some platforms. The testcase fixture does not check if the file exists before making modification to the file, hence causing the error and failure subsequently.

How did you do it?
Add a check to see if the file pmon_daemon_control.json file exists. If the file doesnt exist, then create the file with empty contents. If the file exists, the test continues as is. After the testcase is complete, the file is removed if created by the testcase.

How did you verify/test it?
Ran the testcase under 2 conditions.

When the file is present already. (Passed)
When the file is not present. (Passed) - Checked to see if the file is created and then deleted in the end.