[configlet] Remove transciver check (#9855)
Summary: Remove transceiver check in add rack scenario.

What is the motivation for this PR?
The transceiver data such as rx_power and tx_power is different as expected.
For example, the rxpower before test is
{'rx1power': '0.547', 'rx2power': 'N/A', 'rx3power': 'N/A', 'rx4power': 'N/A', 'temperature': '24.652', 'tx1bias': '7.106', 'tx1power': '0.617', 'tx2bias': 'N/A', 'tx2power': 'N/A', 'tx3bias': 'N/A', 'tx3power': 'N/A', 'tx4bias': 'N/A', 'tx4power': 'N/A', 'voltage': '3.32'}
which is different with
{'rx1power': '0.549', 'rx2power': 'N/A', 'rx3power': 'N/A', 'rx4power': 'N/A', 'temperature': '24.0', 'tx1bias': '7.014', 'tx1power': '0.609', 'tx2bias': 'N/A', 'tx2power': 'N/A', 'tx3bias': 'N/A', 'tx3power': 'N/A', 'tx4bias': 'N/A', 'tx4power': 'N/A', 'voltage': '3.32'}

How did you do it?
Remove the transceiver check.

How did you verify/test it?
E2E test

Signed-off-by: yenlu <yenlu@cisco.com>