Conditionally install env-python3 in setup-container.sh (#9987)
The sonic-mgmt docker image has been updated to globally install all the
packages in env-python3. Now, in the global python3 environment, we have
all the required packages for running tests. Then it is unnecessary to have
the env-python3 virtual environment. Before "env-python3" is deprecated
in sonic-mgmt docker image, we can skip creating "env-python3" in the
setup-container.sh script which is used for setting up sonic-mgmt container
for user different than "AzDevOps" to solve permission issues of mapped
directory.

This change added code to check if pytest is globally installed in python3
environment. If yes, then we can assume that all the env-python3 packages
are also globally installed. Then we can skip creating env-python3 in
setup-container.sh script.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>