Fix test_route_flow_counter (#9958)
What is the motivation for this PR?
When test_route_flow_counter is ran on unsupported platform, it fails due to clear_route_flow_counter fixture running which assumes that platform is supported.

Independent pytest fixtures can run even if one causes the test function to skip. The clear_route_flow_counter fixture should not run if the test is skipped (due to lack of support) so the appropriate fixture arg should be included to mark the dependency.

How did you do it?
Added skip_if_not_supported as a dependency of clear_route_flow_counter. By doing this, if test is skipped because there is a lack of support then clear_route_flow_counter will not run.

How did you verify/test it?
Verified that test is skipped on unsupported platform

Signed-off-by: yenlu <yenlu@cisco.com>