Added verify routes method on vnet vxlan tests. (#9932)
Summary:
Created a method to verify routes configuration on vnet vxlan tests

Details:
This method queries the routes from ASIC DB and counts them, to check that a sufficient amount of routes were added or cleaned.
The method will fail the test in case not enough routes were added or cleaned from the ASIC DB.
The PR also sets a dynamic timeout of route configuration based on the number of routes added or cleaned.

Signed-off-by: yenlu <yenlu@cisco.com>