Add the ability to download a SONiC image for use as the neighbor image
When the neighbor devices are specified to be the type vsonic, allow
downloading a SONiC image if the file doesn't exist already. This
matches the behavior that is present with veos, but requires the user to
specify a file name.

This can (indirectly) be used to also have different SONiC images used
for different testbeds.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>