fix for test_lag_db_status run with dut_lag param as unknown (#9894)
When test_lag_db_status testcase is run with the dut_lag parameter is "unknown" (test_lag_db_status[unknown|unknown]), the testcase fails with the following error: KeyError: 'unknown', at the time it tries to access the namespace_id value from the lag_facts dictionary with the lag name as the key. The testcase checks that if the dut_lag parameter is "unknown", the test should run for all the portchannels. But at this point, this check is after the testcase tries to access the namespace_id value based on the lag name, leading to failure when the dut_lag parameter is "unknown".
Fix: The testcase should first get the list of portchannels if the dut_lag is "unknown", and then fetch the namespace_id for the portchannels.
Need to merge with 202205 manually since there were merge conflicts.