Update function `posix_shell` in class `BaseConsoleConn` to support ascii output.  (#10090)
Description of PR
To support getting output when doing power cycle, we enhance the function posix_shell in class BaseConsoleConn to support ascii output.

What is the motivation for this PR?
To support getting output when doing power cycle, we enhance the function posix_shell in class BaseConsoleConn to support ascii output.

How did you do it?
Set the type of output using sys.setdefaultencoding('utf8').

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>