Raise exceptions from the VM threads as failures in warm-reboot
If there's some exception that happens from the VM thread, raise this as
a failure in the overall warm reboot. This has two purposes: to make any
failures more clearer/easier to find when debugging, and to make sure
that the test fails if something goes wrong there, even if the warm
reboot itself succeeded.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>