Update PORT_READY regex check to match updatePortOperStatus up to up
On Broadcom platforms, at least, after a warm reboot, for physical
interfaces, from updatePortOperStatus, there are only messages about the
port going from up to up. This means that the existing regex check
looking for down to up won't match.

The problem with just changing the regex is that there's a separate
check for the number of interfaces that have flapped, to make sure there
aren't excess flaps. If up to up is matched, then this is counted as a
flap.

Therefore, create a separate state "Start-changes-only" that matches
only down to up, but still keeps the link flap check working correctly.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

Fix regex

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>