[202205] Cherry-pick the PR :Improve the stability of dualtor memory leak test (#9198) (#9836)
This is the manual cherry-pick of PR (#9198) to 202205 branch for conflict resolving.

What is the motivation for this PR?
To improve the stability of dualtor memory leak test.

How did you do it?
Increase the memory threshold buffer from 0.02 to 0.03.
We found that the test occasionally fails on Nvidia 4600C platform due to the memory usage after the test is 103% of the original memory usage, which is 1% more than the buffer defined in the test. I have tried to modify the test and repeat the neighbor leaning action much more times than it is in the test, and no real memory leak was observed.
Since the 0.02 is an experience value, I suggest we increase it a little bit to 0.03.

Get the original memory usage after all neighbors are deleted.
The original memory usage was recorded before deleting the neighbors, suggest doing it after the deleting to avoid any unexpected memory change during the process.

Add retries in the memory check at the end of the test.
We have observed that sometimes after the last memory check in the test, the memory usage could still fall a little without any actions being taken on the switch.