[Snappi] Update test PFC pause frame functionality to include DUT counter polling (#9847)
Summary: Currently, all verification for PFC pause frames are done by polling the ixia device at the end of the test. We need to add functionality to poll the appropriate counters on the DUT while the DUT continues to receive PFC pause frames from the neighboring device.

This PR is dependent on #9775.

Fixes (issue) #9843

Signed-off-by: yenlu <yenlu@cisco.com>