Fix a wrong variable name in TunnelDscpToPgMapping test case (#10063)
Summary:
The variable "dscp" in the debug print should be "inner_dscp":

The print is meant to log the inner_dscp value here.
The variable "dscp" is not assigned beforehand if the dut is Nvidia, so the test will fail.

What is the motivation for this PR?
Fix a wrong variable name in TunnelDscpToPgMapping test case.

How did you do it?
Change it to "inner_dscp"

How did you verify/test it?
Run the test on 4600C test bed, the info was correctly printed.

Signed-off-by: yenlu <yenlu@cisco.com>