Fix Python 3 syntax errors in sonic.py (#10107)
* Fix Python 3 syntax errors in sonic.py

This file is used with the advanced-reboot test case with SONiC
neighbors.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Raise exceptions from the VM threads as failures in warm-reboot

If there's some exception that happens from the VM thread, raise this as
a failure in the overall warm reboot. This has two purposes: to make any
failures more clearer/easier to find when debugging, and to make sure
that the test fails if something goes wrong there, even if the warm
reboot itself succeeded.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Remove whitespace

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Log output on stderr as well

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* For SONiC neighbors, in neigh_lag_status_check, use existing info in cli_info

In the neigh_lag_status_check function, when the neighbor device is
checked to see if there were any LAG flaps, instead of trying to get
this information from the neighbor VM, just use the existing info we
have in self.cli_info already. Otherwise, LAG flaps that may have
happened prior to the test starting would show up here.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Read stdout first, before reading stderr, since the buffer might be full

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>