Fix mux-simulator script not starting due to Werkzeug update (#10229)
Starting on/around September 29th, the mux-simulator script on the host
server failed to start. This is because both the Flask and Werkzeug
python packages updated to version 3.0.0 on that day, and there were
breaking changes in Werkzeug 3.0.0. Our yaml script installs a fixed
Flask version of 2.0.3, and the dependency list for that package only
specified "Werkzeug >= 2.0". This means that pip will happily install
version 3.0.0 even though it is incompatible with this version of Flask.

As a fix, install specifically Flask 2.3.3 and Werkzeug 2.3.7. In
addtion, to make it easier to notice failures here, check to see that
the systemd service is actually running after it gets started.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>