[Fix] Fix get pr_id is None for some repos PR test azure pipelines (#10134)
What is the motivation for this PR?
For sonic-mgmt and sonic-buildimage repo, in PR test azure pipelines, the pr_id is got from the ENV "SYSTEM_PULLREQUEST_PULLREQUESTNUMBER". But for other repos, the ENV name is different.

How did you do it?
If "SYSTEM_PULLREQUEST_PULLREQUESTNUMBER" exists, use it as pr_id, else get from another ENV name.

Signed-off-by: Chun'ang Li <chunangli@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>