Fixed log handling issues on fast/warm upgrade test (#10162)
Fix reboot type source fetch logic: Previously, the reboot type source was based on the test name. However, it doesn't work with the upgrade path test, as it does not include the reboot type in the name of the test. This fix treats the mentioned case by getting the name from the upgrade type parameter.

Signed-off-by: yenlu <yenlu@cisco.com>