Protect vs ValueError when casting to int in conditional_mark (#10298)
* Protect vs ValueError when casting to int in conditional_mark

build_version could be like master.362591-ce7145475 which raises
exceptions when evaluating 5 conditions. This change simply adds
a check if the value is digit before casting it. It also changes
the logger level for this case to ease debugging.

This issue can be reproduced when running a Pytest as described
in KVM Testbed Setup [1] (Please note that test_pretest.py and
test_posttest.py are directly called in Gates)

[1] https://github.com/sonic-net/sonic-mgmt/blob/master/docs/testbed/README.testbed.VsSetup.md

dict_keys object is not subscriptable then it simply
converts it to a list before accessing an index.

Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>

---------

Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>