[syslog/test_logrotate]: Fixed test_logrotate testcase failing due to incorrect logrotate threshold (#9921)
What is the motivation for this PR?
test_logrotate_normal_size testcase fails when the threshold for log rotation is different for different platforms. The threshold for logrotation is set in sonic-buildimage/files/image_config/logrotate/rsyslog.j2. Here the logrotate threshold is set based on the memory available in the platform. In the testcase this threshold was a constant and did not depend on the memory available in the platform.

How did you do it?
Added a function get_threshold_based_on_memory() to first check for the available memory in the platform and then set the logrotate threshold.

How did you verify/test it?
Ran both test_logrotate_normal_size() and test_logrotate_small_size() to see if the testcases passed.