When recovering a DUT with config_reload, wait for BGP sessions to be up
If a DUT is in an unhealthy state, and `config reload` needs to be done
to recover it, then also wait for BGP sessions to come up. This
shouldn't add much delay in most cases. However, there was one case
where during recovery, orchagent hit a crash which brought down most of
the container.

The hopes is that with this additional check, in case orchagent happens
to go down again, the testbed health check will wait here before
starting the test case execution, or fail loudly if something is still
unhealthy.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>