Fix failure case number in test result summary incorrect issue (#10485)
What is the motivation for this PR?
In test summary, failures += failure + error, but in fact, if a testcase failed with both failure and error, the count of failure and error would both +1, so it's unnecessary to calculate in that way

How did you do it?
Let failures += failure

How did you verify/test it?
Run a test with both failure and error, and check the test result xml file

Signed-off-by: yenlu <yenlu@cisco.com>