dualtor_test_tor_ecn mixed and a-a topo failure fix (#10104)
Summary:
When test_dscp_to_queue_during_encap_on_standby and test_ecn_during_encap_on_standby are run in mixed or A-A topology, while randomly selecting PTF IP to test we sometimes select IP mapped to a port with active-active config.. If we select such PTF IP, we won't see expected behavior. With this PR we have added a check to force A-S if this test case is run.

Second part of the change in related to handling the code flow for test cases in test_tor_ecn.py if SYSTEM_DEFAULTS config is not present.