[202205 Cherry-pick] Fix a test issue in verify_dut_health.py (#10466)
This is the manual cherry-pick of PR #10076 for conflict resolving.

Fix test issues in verify_dut_health.py, this script is used in the advanced reboot test.
There are two issues:

In check_services(), when the service check fails, it should not directly fail the test by pytest_assert. Instead, it should raise the RebootHealthError so the exception can be handled and added to the test report.
In the verify_dut_health(), the logic in verify_dut_health before the yield is not correct, the test does not fail on any health errors because any string(failure reason) is considered as True in the all() method.