Update the fwutil test case (#10250)
What is the motivation for this PR?
In the existing implement of the fwutil test, the data(--fw-pkg) is saved in the xx.tar.gz file, and the json file is included in the tar.gz file. Enhance the test to support input a json file to --fw-pkg.

How did you do it?
User can either specify the tar.gz or json file to the --fw-pkg when run the test_fwutil.py

How did you verify/test it?
Run the test_fwutil.py, and test cases pass

Signed-off-by: yenlu <yenlu@cisco.com>