Skip test_read_mac_metadata if required parameters are missing (#10095)
The test arguments image1, image2 and iteration are set to required arguments in test module test_read_mac_metadata.py . If the test pipeline doesn't set the argument, an error is thrown

ERROR: usage: __main__.py [options] [file_or_dir] [file_or_dir] [...]
__main__.py: error: the following arguments are required: --iteration, --image1, --image2
To reduce noise, the arguments are changed to be optional, and the test will be skipped if any of the arguments is missing.

What is the motivation for this PR?
This PR is to skip test_read_mac_metadata if required parameters are missing.

How did you do it?
To reduce noise, the arguments are changed to be optional, and the test will be skipped if any of the arguments is missing.

How did you verify/test it?
The change is verified on a physical testbed. The test is skipped if the argument is not set.

collected 1 item

read_mac/test_read_mac_metadata.py::test_read_mac_metadata[str-msn2700-22] SKIPPED (Skip test due to missing --image1 or --image2)                                                              [100%]
Any platform specific information?
No.

Supported testbed topology if it's a new test case?
Not a new test case.

Signed-off-by: yenlu <yenlu@cisco.com>