[202205] test_ro_disk: Enhancement for reboot of supervisor card (#10528)
The Test Case tacacs/test_ro_disk.py::test_ro_disk, in finally block a reboot is performed, currently, the condition for reboot to be successful is to check the docker is running, the enhancement done is to wait till all the critical process in the docker are up and the interfaces are also up for all the line cards in case the supervisor is rebooted
The Particular improvement is required due to the changes with respect to PR sonic-net/sonic-buildimage#16213. The Parent PR on Master was merged #9893

What is the motivation for this PR?
The Parent Master PR #9893
The Particular improvement is required due to the changes with respect to PR sonic-net/sonic-buildimage#16213.

How did you do it?
Assertion for successful reboot of supervisor card is updated to check all the process are up on critical docker and interfaces are also up on line cards.

How did you verify/test it?
Ran the test cases against a multi-asic line card in a T2 chassis.