enhance restore_bgp_suppress_fib to check status first (#10612)
What is the motivation for this PR?
In current test, it will set the bgp suppress to disabled no matter what's the value in the config_db. If it is enabled before the test, it should honor it, and don't change it after the test.
In this PR, will add a check logic to restore only it is needed.

How did you do it?
To read back the config db before test, then do the restore when it is needed.

How did you verify/test it?
enable the fib suppress.
run bgp suppress fib test.
check the config, if it is same as before.

Signed-off-by: yenlu <yenlu@cisco.com>