Ignore NTP exiting "error" message
When exiting, ntpd from NTPsec logs a message containing ERR at NOTICE
level saying that it is exiting because of a signal. This causes
loganalyzer to flag the message because ERR is present. This log message
happens during graceful termination (such as when systemd is restarting
the service). Therefore, ignore it.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>