Merged PR 3970: Fix fanout unreachable issue after ansible is upgraded
Fix fanout unreachable issue after ansible is upgraded

We recently upgraded the Ansible version to 6.7.0 in the sonic-mgmt docker image.
If run tests using python3 and the latest sonic-mgmt docker, creating the
fanouthosts fixture will fail with a “Host Unreachable” error if the fanout
switch is running SONiC.

The reason for this is that we supplied both “ansible_ssh_pass” and
“ansible_ssh_password” variables while initializing the SONiC fanouthosts fixture.

After the Ansible version was upgraded, the order of using the variables might
have changed, and an inconsistent value of this variable caused an authentication
issue, which eventually resulted in the “Host Unreachable” error.

Fanout hosts are defined under the “fanout” group in the inventory file.
The group vars file ansible/fanout/secrets.yml defines “ansible_ssh_password” as
follows:

    ansible_ssh_password: “{{ secret_group_vars[‘fanout’][‘ansible_ssh_pass’] }}”

Based on the current secrets.json, “ansible_ssh_password” is rendered to “fanoutpassword”.

While initializing SONiC fanouthosts, line #76 in tests/common/devices/sonic.py tries to
initialize the variable “ansible_ssh_pass” to the current value “password”.

Now we have both:
    ansible_ssh_password: “fanoutpassword”
    ansible_ssh_pass: “password”

With the older Ansible version, “ansible_ssh_pass” is used, and it works for SONiC-based
fanout switches. With the latest Ansible version, “ansible_ssh_password” is used and
results in an authentication failure.

The fix is to update the variable name in “fanout/secrets.yml” to “ansible_ssh_pass”.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>