[test_qos_sai] Start the mux on the upper ToR in teardown (#10656)
Approach
What is the motivation for this PR?
Fix the issue that the testcase that runs after test_qos_sai fails to toggle to the upper ToR.
The reason is that, in the teardown of test_qos_sai, it doesn't start the mux container on the upper ToR, and the mux feature disable config is saved to config_db.json by enable_container_autorestart.

Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
Start the mux in teardown.

How did you verify/test it?

Signed-off-by: yenlu <yenlu@cisco.com>