Fix TypeError in check_sysfs (#10702)
What is the motivation for this PR?
This PR is to fix TypeError in check_sysfs. The error message is as below.

How did you do it?
Convert sfp_id into int explicitly.

How did you verify/test it?
The change is verified by running test_cold_reboot on a Mellanox platform.