Fix an error in parsing the neighbor analysis for SONiC neighbors (#10685)
In #10107, when doing a warm reboot with SONiC neighbors, instead of
going to the neighbor device to get the LAG flap information, the code
was changed to reuse information that was already collected and
returned. However, the code to access that information is wrong, and
needs to access the `po` key first.

Fix this error by updating the code.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>