Gracefully handle hosts unreachable for AnsibleHostBase of devutil (#10639)
What is the motivation for this PR?
The AnsibleHostBase class of devutil is not able to distinguish host unreachable scenario and run ansible module failed scenario.

How did you do it?
This change enhanced the AnsibleHostBase class to explicitly raise HostsUnreachable exception when any of the host is unreachable while running ansible module.

This change also added a "reachable" method to the AnsibleHostBase class. This method runs ansible "ping" method to check if SSH connection can be successfully established to the hosts. It returns the test results in a tuple. The first in tuple is a bool explicitly indicating if all the hosts are reachable or not. The second item is a dict of detailed ping test results.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>