Fix test_pfcwd_timer_accuracy flakiness
Address flakiness in this test, which mainly comes from two problems: detecting PFC storm, and
asserting that the time taken to detect a PFC storm occurs very shortly after the storm is
triggered (within 800ms). This rarely occurs on our duts; the time is probably closer to around 5
seconds on average.

The PFC storm detection has a chance to randomly fail but it's not clear yet how to reliably
reproduce this event. As a workaround, added retry logic to the test to retrigger storm if
detection fails.
Consolidate the t2 and non-t2 checks to be based around asserting that time between a
storm's detection and restoration is within the expected time window of when the test starts and
stops a storm.