Porting PR:9758 to 202305 - qos/test_qos_sai.py changes from master to 202305 (#10722)
Porting PR:9758 to 202305.

Changes include:

A fixture to add static route via the backplane port-channels for dst-IP addresses in the SRC Asic for T2 topology and chassis-packet type of routers.
Remove LossyQueueVoq second case - which tests only default voq. This will be revisited later.
Add a skip-fixture for testQosSaiSharedResSize if the number of ports is not enough.
Add the parameter: pkts_num_egr_mem for cisco platforms as well.
Skip losslessVoq for mutli-asic.
Add dynamic calculation of dscpToQueue mapping for allPortQueueWatermark test instead of the hard-coded values. Also add support for multi-asic to the testcase.