Fix testQosSaiDwrr (#10764)
Because the method to generate congestion for our platform is changed from disable the whole port to disable the queue one by one, and when we recover the port, we also enable the queue one by one. So, it will cause testQosSaiDwrr to fail. testQosSaiDwrr is to test the function of scheduling queue to send packets, the pre-condition
is that all queues should be at the same enabled state, but currently they are enabled back one by one, so the order to send packets is the same as the order to enable the queue not depending on the scheduling algorithm. So, for this case, we still use the original method to generate congestion.

Change-Id: Idba01127df9c55a79cb5ca513d1c552ae1c33f30