[dualtor] Fix toggle request retry failure
Fix the error observed in nightly:
TypeError("__init__() got an unexpected keyword argument 'allowed_methods'")

This is due to that the urllib3 used in nightly agents is of version
'1.25.8', which have argument `method_whitelist` instead of
`allowed_methods` to take the allowed HTTP methods.

Let's check first before init `Retry`.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>