[dualtor] Fix toggle request retry failure (#10863)
Approach
What is the motivation for this PR?
Fix the error observed in nightly:
TypeError("init() got an unexpected keyword argument 'allowed_methods'")

It is introduced by PR: #10739

This is due to that the urllib3 used in nightly agents is of version
'1.25.8', which have argument method_whitelist instead of
allowed_methods to take the allowed HTTP methods.

Let's check first before init Retry.

Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
As the motivation.

How did you verify/test it?

Signed-off-by: yenlu <yenlu@cisco.com>