Fix test_counterpoll_watermark flakiness
This test randomly picks a counter to test out of 3.
One check is only meant to run on one of these counter but ends up
running for all counters.
There is another call for this check that is properly gated a few lines
above, this fix essentially make sure it works the same.

The check actually works for all counters per my investigation but the
delay is not sufficient for some older products. These take longer to
properly populate the WATERMARK data in COUNTERS_DB.