[dualtor-io] Increase stop wait to 90s (#10911)
The reason is that, on dualtor-120 testbed, dual_tor_sniffer sometimes takes around 45+ seconds to be stopped after receiving the SIGINT. This is because dual_tor_sniffer script captures more packets on dualtor-120 as it has more vlan members and dualtor-io tests are sending more packets. After 30 seconds, supervisorctl will kill the dual_tor_sniffer, which will result in the loss of packets.
Let's increase the stopwaitsecs to 90s.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>