[dualtor] Fix mux config manual leftover (#10924)
Approach
What is the motivation for this PR?
Same purpose as PR: #10555.
This is to ensure mux auto to save to config_db.json to avoid any unwanted modification.

How did you do it?
Do a config save in teardown.

How did you verify/test it?
Any platform specific information?
Supported testbed topology if it's a new test case?

Signed-off-by: yenlu <yenlu@cisco.com>