[sanity_check] Use golden config to reboot for `adaptive_recover` (#10943)
What is the motivation for this PR?
For adaptive_recover, it will reboot if monit has failures. But the DUT keeps using the config_db.json, which might be corrupted by previous tests in nightly.
Let's reboot with the running golden config for the reboot in adaptive_recover.

How did you do it?
Override config_db.json with the running golden config before reboot.

How did you verify/test it?
Run on dualtor testbed with monit service routeCheck fails and config_db.json is corrupted with mux config manual.
Verify after reboot, monit is recovered and mux config is back to auto.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>